Split location model class file
